Fix integer map value JSON generation
Integer map values were previously output as with format %@, which
caused a crash because the parameter was shortValue. This seems to be an
existing bug, but no FE thrift uses map<k, int> at runtime, so it hasn't
affected anything. List and Set seem to work, so special-case Map to
always use primitive format instead of id format.